+++ app/src/main/java/com/raveendra/finalproject_binar/data/repository/CourseRepository.kt
@@ -1,18 +1,18 @@
 package com.raveendra.finalproject_binar.data.repository
 
 
-import com.raveendra.finalproject_binar.data.dummy.CourseDataSource
+import com.raveendra.finalproject_binar.data.dummy.DummyCourseDataSource
 import com.raveendra.finalproject_binar.model.Course
 
 interface CourseRepository {
-    fun getProducts(): List<Course>
+    fun getCourses(): List<Course>
 }
 
 class CourseRepositoryImpl(
-    private val productDataSource: CourseDataSource
+    private val courseDataSource: DummyCourseDataSource
 ) : CourseRepository {
 
-    override fun getProducts(): List<Course> {
-        return productDataSource.getCourseList()
+    override fun getCourses(): List<Course> {
+        return courseDataSource.getCourseList()
     }
 }
