+++ app/src/main/java/com/raveendra/finalproject_binar/presentation/course/adapter/CourseAdapter.kt
@@ -5,11 +5,12 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.AsyncListDiffer
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
-import coil.load
 import com.raveendra.finalproject_binar.core.ViewHolderBinder
 import com.raveendra.finalproject_binar.databinding.ItemListCourseFreeBinding
 import com.raveendra.finalproject_binar.databinding.ItemListCoursePremiumBinding
 import com.raveendra.finalproject_binar.model.Course
+import com.raveendra.finalproject_binar.presentation.course.viewholder.CourseFreeViewHolder
+import com.raveendra.finalproject_binar.presentation.course.viewholder.CoursePremiumViewHolder
 
 class CourseAdapter(
     var courseTypeAdapter: CourseTypeAdapter,
@@ -59,48 +60,6 @@ class CourseAdapter(
 
     override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         (holder as ViewHolderBinder<Course>).bind(differ.currentList[position])
-    }
-
-    class CoursePremiumViewHolder(
-        private val binding: ItemListCoursePremiumBinding,
-        private val onClickListener: (Course) -> Unit
-    ) : RecyclerView.ViewHolder(binding.root), ViewHolderBinder<Course> {
-        override fun bind(item: Course) {
-            binding.ivImg.load(item.imgUrl) {
-                crossfade(true)
-            }
-            binding.tvTitle.text = item.name
-            binding.tvAuthor.text = item.author
-            binding.tvRating.text = item.rating.toString()
-            binding.tvTitle.text = item.name
-            binding.tvLevel.text = item.level
-            binding.tvModul.text = item.modul
-            binding.tvDuration.text = item.duration
-            binding.root.setOnClickListener {
-                onClickListener.invoke(item)
-            }
-        }
-    }
-
-    class CourseFreeViewHolder(
-        private val binding: ItemListCourseFreeBinding,
-        private val onClickListener: (Course) -> Unit
-    ) : RecyclerView.ViewHolder(binding.root), ViewHolderBinder<Course> {
-        override fun bind(item: Course) {
-            binding.ivImg.load(item.imgUrl) {
-                crossfade(true)
-            }
-            binding.tvTitle.text = item.name
-            binding.tvAuthor.text = item.author
-            binding.tvRating.text = item.rating.toString()
-            binding.tvTitle.text = item.name
-            binding.tvLevel.text = item.level
-            binding.tvModul.text = item.modul
-            binding.tvDuration.text = item.duration
-            binding.root.setOnClickListener {
-                onClickListener.invoke(item)
-            }
-        }
     }
